+++ src/components/journal/field-bottom-sheet.tsx
@@ -17,7 +17,7 @@ import {
 } from "@expo/ui/swift-ui/modifiers";
 
 import { FIELD_ICONS, FIELD_LABELS, FieldType } from "@/core/constants";
-import { FIELD_TYPES } from "@/hooks/journal/use-journal-field";
+import { FIELD_TYPES } from "@/utils/journal/use-journal-field";
 
 type Props = {
   /** ボトムシートの表示状態 */
